Add files via upload
In this prject I tried to create a e-commerce website based on mobile phones. 

Site design is fully responsive and clear. There are some features like - zoom, hover effects - give feedback to users while surfing on the site. 
Besides, there are some calculation layers from taxes to quantity in the back field.

Site has three main pages which are main page that shows the phone. Second one is details page, shows some features of phone when clicked the images. And last one is payment page that display the payment details to users.